CategoriesList added to Categories page
